update to version 0.0.4
- Correction: The cost of $2.5 trillion to transition to 80\% wind and
solar energy was previously stated as the cost for California alone.
This is in fact the projected cost for all of the United States. This
error was due to an initial mis-reading of the technologyreview.com
article: "The $2.5 trillion reason we can't rely on batteries to clean
up the grid".
- Other minor updates.